commit 010

+++ gestion-évenement/database/migrations/2023_05_29_202901_create_categories_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('categories', function (Blueprint $table) {
-            $table->increments('id_category');
+            $table->unsignedBigInteger('id_category');
             $table->string('name_category');
             $table->timestamps();
         });
